requirements.txt
requests==2.21.0
six==1.12.0
websocket==0.2.1
websocket-client==0.54.0
PyYAML==5.1
urllib3==1.23
certifi==2018.4.16
chardet==3.0.4
idna==2.7